A6: split AppointmentRepository into Reader + Writer behind a façade (#563 backlog)
Instance-based variant of the read/write split. AppointmentReader and
AppointmentWriter each extend AbstractRepository (reusing the same wpdb binding,
table name, cache group and inherited helpers); the façade composes both and
delegates the 29 domain methods. The generic CRUD inherited from
AbstractRepository (findById/insert/update/delete/…) stays on the façade so
callers using it directly are unaffected. Because all three instances bind the
same global $wpdb (one connection), transactions and FOR UPDATE locks remain
coherent across delegated reads/writes. Tech-debt note points to the future
call-site migration (#563 B3).

Behavior-preserving — AppointmentRepositoryTest + self-scheduling dependents
green (431 tests); PHPStan L8 + WPCS clean.
